Make samplers and `While` loops `opaque` (#100)
Reason: For while loops, the definition is not very helpful, it's better
to rely on lemmas around it for proofs. When calling samplers, we should
also not rely on their internals, but on their correctness properties.
It also reduces resource usage.

As a concession, this PR also adds a postcondition for `While` to avoid
having to call a lemma to handle the case of a while loop immediately
exiting.

By submitting this pull request, I confirm that my contribution is made
under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).