Model nontermination explicitly ("partial Hurd monad") (#104)
This PR changes the type of `Monad.Result<A>` from `Result(value: A,
rest: Rand.Bitstream)` to `Diverging | Result(value: A, rest:
Rand.Bitstream)`. In order to make it easier to work with, it also adds
a few combinators for `Monad.Result` in order to avoid pattern matching
everywhere.

The main benefit of this is that `While` loops can be properly defined
without `assume false` and we can prove Hurd's theorem 38 (while loop
unrolling) in Dafny. Another benefit is that we can reason about
nontermination probabilities in Dafny via `Rand.prob(iset s |
Sample(s).Diverging?)` for a probabilistic computation `Sample:
Hurd<A>`.

For reasoning about the measurability of results, I also introduced some
infrastructure to lift measure spaces on `A` to `Result<A>`, see
`Monad.ResultEventSpace`, `Monad.LiftInEventSpaceToResultEventSpace`,
and `Monad.LiftRestInEventSpaceToResultEventSpace`. In fact, the
measurability proofs required the most modifications as a result of the
change.

By submitting this pull request, I confirm that my contribution is made
under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).