Separate modules for equivalence proofs (#108)
Moving all the code that is referred to in `Implementation` to establish
the equivalence between the imperative and functional model into a
separate module `Equivalence`. The code in the new module should not be
referred to by any other module than `Implementation`.

By submitting this pull request, I confirm that my contribution is made
under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).