Prove correctness of the while loop in BernoulliExpNeg (#129)
This PR `assume`s measurability of sets and independence of values of
type `Hurd`. (Proving these is not hard, but tedious, so I'm deferring
this to later.)

By submitting this pull request, I confirm that my contribution is made
under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).