Add more efficient Int32 Shuffling (#168)
Replaces the previous shuffle method with a version that operates on
`int32` instead of `int` in Dafny, that is, `int` instead of
`BigInteger` in Java. Tests have shown that this leads to a speed-up as
it avoids previously present conversions between `BigInteger` and `int`.
The new method lifts a newly required external method `UniformSample32`,
which operates on `int32` (in Dafny). All other methods (that is,
samplers) keep lifting `UniformPowerOfTwoSample`, which operates on
`int` (in Dafny).

By submitting this pull request, I confirm that my contribution is made
under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).